feat: UserIdMapTask rebuilt on student_anonymoususerid
This mapping used to be produced by simply hashing the user ID with a
common hashing function (MD5) which happens to be supported by MySQL.
Also, the mapping used to be fully deterministic using the hashing
function, user_id, and secret_key alone.  Now, both of the above points
are no longer true as of >1 year ago:

1. The hashing function used in production has switched from MD5 to
   something else, and that new function is not built into MySQL.
2. The mapping is no longer deterministic because the edx-platform logic
   was changed to memorize (cache) old mappings in a table
   (student_anonymoususerid).

Fortunately, we no longer need to actually calculate the
anonymous_user_id anymore because the cache table is easily queried.
This is good because it simplifies the pipeline logic, and reduces the
number of reasons why we need access to the SECRET_KEY.

DENG-1356